add a config command, fix reading camelcase configs
The config command helps see what config settings are currently in
place. This is non-trivial, since we have layers including defaults, the
ini file, and environment variables. It's also useful for the tests.

And, as spotted by @Mic92, support reading configs where the keys are in
any casing the user wants. apiURL is good for readability, for example.
We had missed that the ini library now lowercases keys by default, since
the tests didn't cover that before.