adapt to 2fa login api breakage
The twoFactorProviders response field is now a list of strings, not a
list of integers. I'm not sure if this was an OK change to make because
the languages upstream uses don't care about ints versus ints inside
strings, but Go's json library certainly does.

At least we can treat the strings as being entirely opaque, though.
After this small change, 'go test' passes again.