add support for multiple two-factor auth providers
If more than one provider is enabled for the account, an extra prompt is
shown, leading the user to choose which provider they want to use.

The user selects a number from 1 to N from the available options, which
is not the same integer as the internal enum for each provider. This
helps, because making the user choose between 1-2-3 is easier than
0-2-5, for example. The internal enum is also an internal detail.

Fixes #12.