decouple the crypto code and its globals into a struct
This allows the state for deriving keys and decoding data to be
controlled separately from direct globals, allowing us to start having
table-based unit tests for decrypting cipher strings.

The crypto funcs are generally converted to methods, coupled to the
state struc, which we've called secretCache.

We add the first test case, which simply decrypts a simple string in the
cipher string format used by vault.bitwarden.com at the time of writing.

For #22.