support encrypted communication over dbus
If the client supports non-plaintext sessions, implement them.

We also stop using cipher names as dbus labels, because those are never
encrypted, whereas Bitwarden cipher names are encrypted. So, when in
doubt, prefer not to leak them.

Otherwise, secret-tool works just as before, since it supports both
session encryption algorithms. The client first tries the encrypted
form, and if the server does not support it, the client falls back to
plaintext.

Fixes #17.