avoid a "dump" panic on card ciphers
Ciphers storing a credit or debit card don't have a "Login" object,
so we would panic in the "dump" command when attempting to print login
fields such as the URL or password.

For now, simply filter ciphers depending on whether Login exists. We
should probably improve this command in the future, but for now at least
it doesn't panic.

While at it, stop using tabwriter for it. First, if one has hundreds of
passwords stored, the command would seem to hang for tens of seconds,
decrypting all secrets before printing the whole aligned table. Now, it
prints each line as it goes. Not aligned anymore, but that's not a
terrible thing either - some URLs can be very long and mess up the output.

We also start obeying ^C properly in both the "dump" command and when
reading a password. Before the fix, both would simply keep going until
they were finished.

Finally, update deps.

Fixes #28.