Add organization support (#44)
Adds basic support for organizations. I tested only dump so far, but
also implemented support for serve.

Each user profile has a private key. It is pkcs#8 encoded RSA. This key
is encrypted with the accounts symmetric master key. The organizations
stored in Profile.Organizations each have an encrypted symmetric master
key, with which the organization ciphers are encrypted (or technically
an encryption key as the first 32 bytes, and a mac key as the last 32
bytes).

The procedure is now as follows: With the master key, decrypt the
private key. Save the private key in the secretCache. For each
organization, decrypt the encrypted organization symmetric master key
using the private key, and store them in a map in the secretCache.

To decrypt a cipher, the code then has to retreive the keys from the
secret cache.

I used decryptWith instead of secrets.decrypt, since the cipherstring
does not know it is encrypted with the organizations key, and thus
secrets.decrypt uses the user's own keys.

Fixes #19.